requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-ckeditor==6.0.0
django-embed-video==1.4.0
django-js-asset==1.2.2
django-robots==4.0
django-sitemaps==1.0
gunicorn==20.0.4
idna==2.10
mailchimp-marketing==3.0.36
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
